id type to long

+++ tanks_backend/src/main/resources/db/migration/V1__initial_schema.sql
@@ -1,7 +1,7 @@
 ﻿CREATE SCHEMA IF NOT EXISTS game;
 
 CREATE TABLE IF NOT EXISTS game.user (
-                                         id integer PRIMARY KEY,
+                                         id bigint PRIMARY KEY,
                                          username varchar UNIQUE NOT NULL,
                                          password_hash varchar NOT NULL,
                                          email varchar UNIQUE NOT NULL,
@@ -13,8 +13,8 @@ CREATE TABLE IF NOT EXISTS game.user (
 );
 
 CREATE TABLE IF NOT EXISTS game.leaderboard (
-                                        id integer PRIMARY KEY,
-                                        user_id integer UNIQUE,
+                                                id      bigint PRIMARY KEY,
+                                                user_id bigint UNIQUE,
                                         wins integer DEFAULT 0,
                                         losses integer DEFAULT 0,
                                         kills integer DEFAULT 0,
@@ -25,18 +25,18 @@ CREATE TABLE IF NOT EXISTS game.leaderboard (
 );
 
 CREATE TABLE IF NOT EXISTS game.team (
-                                         id integer PRIMARY KEY,
+                                         id bigint PRIMARY KEY,
                                          name varchar UNIQUE NOT NULL,
                                          created_at timestamp,
                                          updated_at timestamp
 );
 
 CREATE TABLE IF NOT EXISTS game.game_session (
-                                         id integer PRIMARY KEY,
+                                                 id             bigint PRIMARY KEY,
                                          name          varchar UNIQUE NOT NULL,
                                          password      varchar,
                                          status varchar,
-                                         winner_team_id integer,
+                                                 winner_team_id bigint,
                                          start_time timestamp,
                                          end_time timestamp,
                                          game_settings jsonb,
@@ -46,10 +46,10 @@ CREATE TABLE IF NOT EXISTS game.game_session (
 );
 
 CREATE TABLE IF NOT EXISTS game.game_session_player (
-                                        id integer PRIMARY KEY,
-                                        player_id integer,
-                                        game_session_id integer,
-                                        team_id integer,
+                                                        id              bigint PRIMARY KEY,
+                                                        player_id       bigint,
+                                                        game_session_id bigint,
+                                                        team_id         bigint,
                                         kills integer DEFAULT 0,
                                         deaths integer DEFAULT 0,
                                         joined_at timestamp,
@@ -57,8 +57,8 @@ CREATE TABLE IF NOT EXISTS game.game_session_player (
 );
 
 CREATE TABLE IF NOT EXISTS game.game_events (
-                                        id integer PRIMARY KEY,
-                                        game_session_id integer,
+                                                id              bigint PRIMARY KEY,
+                                                game_session_id bigint,
                                         event_type varchar,
                                         event_time timestamp,
                                         summary jsonb
